feat: move form to ui directory
